feat: Add user_id field to contact model and update contact controller to handle user-specific contacts
- Add user_id field to the contact model to associate contacts with users
- Update getAllContacts and createContact functions in the contact controller to only retrieve and create contacts for the authenticated user
- Update updateContact and deleteContact functions in the contact controller to check if the authenticated user has permission to update or delete the contact

Closes #123

+++ controllers/contact.controller.js
@@ -2,7 +2,7 @@ import asyncHandler from 'express-async-handler';// i use instead of try catch
 import { Contact } from '../models/contact.model.js';
 
 export const getAllContacts = asyncHandler(async(req,res)=>{
-   const contacts = await Contact.find();
+   const contacts = await Contact.find({user_id:req.user.id});
    res.status(200).json(contacts)
 })
 
@@ -23,7 +23,7 @@ export const createContact = asyncHandler(async(req,res)=>{
   }
 
   const contact = await Contact.create({
-    name,email,phone,
+    name,email,phone,user_id:req.user.id
   })
   res.status(201).json(contact)
 })
@@ -34,6 +34,12 @@ export const updateContact = asyncHandler(async(req,res)=>{
       res.status(404);
       throw new Error('contact not found')
     }
+
+    if(contact.user_id.toString() !== req.user.id){
+    res.status(403);
+    throw new Error('User dont have permission to update other user contacts')
+    }
+
     const updatedContact = await Contact.findByIdAndUpdate(
       req.params.id,req.body,{new:true},
     )
@@ -46,6 +52,12 @@ export const deleteContact = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error('Contact not found');
   }
+
+  if(contact.user_id.toString() !== req.user.id){
+    res.status(403);
+    throw new Error('User dont have permission to delete other user contacts')
+    }
+
 
   await contact.deleteOne(); // Use deleteOne() method instead of remove()
   res.status(200).json({ Message: 'Contact deleted successfully' });
